Refactor import paths for resetPasswordAction and signOutAction
